refactor: move to typescript
