+++ lib/constants/index.ts
@@ -1,0 +1,4 @@
+export const MAX_DEPOSIT_PLANS = 2;
+export const DEPOSIT_TYPE_ONE_TIME = "One time";
+export const DEPOSIT_TYPE_MONTHLY = "Monthly";
+export const VALID_PLAN_TYPES = [DEPOSIT_TYPE_ONE_TIME, DEPOSIT_TYPE_MONTHLY];
