+++ package.json
@@ -1,8 +1,18 @@
 {
   "scripts": {
+    "start": "ts-node index.js",
     "test": "jest"
   },
   "devDependencies": {
-    "jest": "^27.1.0"
+    "jest": "^27.1.0",
+    "ts-node": "^10.2.1",
+    "ts-node-dev": "^1.1.8",
+    "typescript": "^4.4.2"
+  },
+  "dependencies": {
+    "@babel/core": "^7.15.4",
+    "@babel/preset-env": "^7.15.4",
+    "@babel/preset-typescript": "^7.15.0",
+    "babel-jest": "^27.1.0"
   }
 }
